requirements.txt
mcp-python>=0.1.0
httpx>=0.24.0
